Move S3 client building into a util
- Rename `TestingConstants` to `TestngUtil`
- Add util method to start an S3 builder
