+++ trino-s3-proxy/src/test/java/io/trino/s3/proxy/server/TestProxiedAssumedRoleRequests.java
@@ -18,8 +18,8 @@ import io.airlift.http.server.testing.TestingHttpServer;
 import io.trino.s3.proxy.server.credentials.Credentials;
 import io.trino.s3.proxy.server.rest.TrinoS3ProxyRestConstants;
 import io.trino.s3.proxy.server.testing.ManagedS3MockContainer.ForS3MockContainer;
-import io.trino.s3.proxy.server.testing.TestingConstants.ForTesting;
 import io.trino.s3.proxy.server.testing.TestingCredentialsRolesProvider;
+import io.trino.s3.proxy.server.testing.TestingUtil.ForTesting;
 import io.trino.s3.proxy.server.testing.harness.TrinoS3ProxyTest;
 import io.trino.s3.proxy.server.testing.harness.TrinoS3ProxyTestCommonModules.WithConfiguredBuckets;
 import org.junit.jupiter.api.AfterAll;
